backend/requirements.txt
fastapi
uvicorn
gspread
oauth2client
pandas
python-multipart
google-auth
